MPT-10121 Discount level 14 customer got level 1 pricing after order validation
